ENH: added time and scan routines to libraries
Moved commonly used time and radarscan routines to the existing libraries.
