+++ codebase/superdarn/src.bin/tk/tool/make_grid.2.0/make_grid.c
@@ -25,6 +25,7 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 Modifications:
  2021-01-11 Marina Schmidt fixing concatenation bug with start/end time, and deprecated -c flag
+ 2021-08-27 Angeline G Burrell removed duplicate code now present in libraries
 
 */
 
@@ -96,50 +97,6 @@ double minsrng=-1;
 double maxsrng=-1;
 
 struct GridTable *grid;
-
-
-
-/**
- *
- **/
-int exclude_outofscan(struct RadarScan *ptr) {
-
-    int n,num=0;
-    struct RadarBeam *tmp;
-
-    if (ptr==NULL) return -1;
-    if (ptr->num==0) return -1;
-    tmp=malloc(sizeof(struct RadarBeam)*ptr->num);
-    if (tmp==NULL) return -1;
-
-    for (n=0;n<ptr->num;n++) {
-
-        if (ptr->bm[n].scan<0) continue;
-
-        memcpy(&tmp[num],&ptr->bm[n],sizeof(struct RadarBeam));
-        num++;
-
-    }
-
-    free(ptr->bm);
-
-    if (num>0) {
-        ptr->bm=realloc(tmp,sizeof(struct RadarBeam)*num);
-        if (ptr->bm==NULL) {
-            free(tmp);
-            ptr->num=0;
-            return -1;
-        }
-    } else {
-        free(tmp);
-        ptr->bm=NULL;
-    }
-
-    ptr->num=num;
-  
-    return 0;
-}
-
 
 
 /**
@@ -221,52 +178,6 @@ void parse_ebeam(char *str) {
 
 }
 
-
-
-/**
- * Converts an input date from YYYYMMDD format to an epoch time in number of
- * seconds since 00:00 UT on January 1, 1970.
- **/
-double strdate(char *text) {
-
-    double tme;
-    int val;
-    int yr,mo,dy;
-
-    /* Calculate day, month, and year from YYYYMMDD format date */
-    val=atoi(text);
-    dy=val % 100;
-    mo=(val / 100) % 100;
-    yr=(val / 10000);
-
-    /* If only 2-digit year provided then assume it was pre-2000 */
-    if (yr<1970) yr+=1900;
-
-    /* Calculate epoch time of input year, month, and day */
-    tme=TimeYMDHMSToEpoch(yr,mo,dy,0,0,0);
-
-    /* Return epoch time in number of seconds since 00:00UT on January 1, 1970 */
-    return tme;
-
-}
-
-
-/**
- * Converts an input time from HHMM format to number of seconds.
- **/
-double strtime(char *text) {
-
-    int hr,mn;
-    int i;
-
-    for (i=0;(text[i] !=':') && (text[i] !=0);i++);
-    if (text[i]==0) return atoi(text)*3600L;
-    text[i]=0;
-    hr=atoi(text);
-    mn=atoi(text+i+1);
-    return hr*3600L+mn*60L;
-
-}
 
 int rst_opterr(char *txt) {
     fprintf(stderr,"Option not recognized: %s\n",txt);
